move `LegacyComboCounter` to `osu.Game.Skinning`
