+++ osu.Game.Tests/Visual/Gameplay/TestSceneSkinnableComboCounter.cs
@@ -10,6 +10,7 @@ using osu.Framework.Testing;
 using osu.Game.Rulesets.Osu;
 using osu.Game.Rulesets.Scoring;
 using osu.Game.Screens.Play.HUD;
+using osu.Game.Skinning;
 
 namespace osu.Game.Tests.Visual.Gameplay
 {
